rm old summarize_gather_at
